refactor(topology_cleanup): extract merge_shared_mutation_branches from prune command
- Move merge_shared_mutation_branches and 8 helper functions/structs to representation/algo/topology_cleanup/merge_shared_mutations.rs alongside collapse.rs
- Move tests to topology_cleanup/__tests__/test_merge_shared_mutations.rs
- Update imports in prune, optimize, and their test files to use new module path
