+++ packages/treetime/src/commands/prune/__tests__/mod.rs
@@ -1,2 +1,1 @@
-mod test_merge_shared_mutations;
 mod test_run;
