+++ packages/treetime/src/commands/prune/__tests__/test_run.rs
@@ -1,9 +1,8 @@
 #[cfg(test)]
 mod tests {
   use crate::alphabet::alphabet::Alphabet;
-  use crate::commands::prune::run::{
-    collapse_sparse_edges_from_leaf_recursive, get_edge_num_muts, merge_shared_mutation_branches, prune_nodes,
-  };
+  use crate::commands::prune::run::{collapse_sparse_edges_from_leaf_recursive, get_edge_num_muts, prune_nodes};
+  use crate::representation::algo::topology_cleanup::merge_shared_mutations::merge_shared_mutation_branches;
   use crate::gtr::get_gtr::{JC69Params, jc69};
 
   use crate::representation::partition::marginal_sparse::PartitionMarginalSparse;
