+++ packages/treetime/src/representation/algo/topology_cleanup/__tests__/mod.rs
@@ -1,2 +1,3 @@
 mod test_collapse_edge;
+mod test_merge_shared_mutations;
 mod test_reroot;
