+++ packages/treetime/src/representation/algo/topology_cleanup/__tests__/test_merge_shared_mutations.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests {
   use crate::alphabet::alphabet::Alphabet;
-  use crate::commands::prune::run::merge_shared_mutation_branches;
+  use crate::representation::algo::topology_cleanup::merge_shared_mutations::merge_shared_mutation_branches;
   use crate::gtr::get_gtr::{JC69Params, jc69};
 
   use crate::representation::partition::marginal_sparse::PartitionMarginalSparse;
