+++ packages/treetime/src/representation/algo/topology_cleanup/mod.rs
@@ -2,4 +2,5 @@
 mod __tests__;
 
 pub mod collapse;
+pub mod merge_shared_mutations;
 pub mod reroot;
